requirements.txt
# Dependencias principales
tensorflow>=2.9.0
numpy>=1.21.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
pandas>=1.3.0
seaborn>=0.11.0
pillow>=9.0.0
flask>=2.0.0
requests>=2.25.0

# Soporte para imágenes HEIC/HEIF (opcionales)
pillow-heif>=0.10.0
pyheif>=0.7.0

# Dependencias de desarrollo (opcionales)
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0